# Faster Hash duping in #json_clone