@mperham 2019-10-06
For future optimizations, can you use a real job payload? Micro benchmarks are very sensitive to the exact setup and that could easily make a diff.