@fatkodima 2019-10-06
Sure. This payload is already much looks like real ('cause every payload have `class` as a string, `args` as an array, and `jid` as a string). Wdyt?

Also reordered first 2 branches, because it seems as `String` and `Integer` are the most popular types for job arguments, and when argument is a string it is faster to check first branch (as now) than second after heavy first branch (as before). For integers there would not be much difference. On my benchmark this change gave another 10%.