@mperham 2019-10-06
Real payloads are 2-3x bigger in byte size and have twice as many keys. 

> On Oct 6, 2019, at 10:48, fatkodima <notifications@github.com> wrote:
> 
> ﻿
> Sure. This payload is already much looks like real ('cause every payload have class as a string, args as an array, and jid as a string). Wdyt?
> 
> Also reordered first 2 branches, because it seems as String and Integer are the most popular types for job arguments, and when argument is a string it is faster to check first branch (as now) than second after heavy first branch (as before). For integers there would not be much difference. On my benchmark this change gave another 10%.
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub, or mute the thread.