@fatkodima 2019-10-06
```ruby
# frozen_string_literal: true

require "benchmark/ips"

def deep_dup(obj)
  if Integer === obj || Float === obj || TrueClass === obj || FalseClass === obj || NilClass === obj
    obj
  elsif String === obj
    obj.dup
  elsif Array === obj
    obj.map { |e| deep_dup(e) }
  elsif Hash === obj
    duped = obj.dup
    duped.each_pair do |key, value|
      duped[key] = deep_dup(value)
    end
    duped
  else
    obj.dup
  end
end

def deep_dup_optimized(obj)
  if String === obj
    obj.dup
  elsif Integer === obj || Float === obj || TrueClass === obj || FalseClass === obj || NilClass === obj
    obj
  elsif Array === obj
    obj.map { |e| deep_dup_optimized(e) }
  elsif Hash === obj
    duped = {}
    obj.each_pair do |key, value|
      duped[key] = deep_dup_optimized(value)
    end
    duped
  else
    obj.dup
  end
end

def deep_dup_original_branches(obj)
  if Integer === obj || Float === obj || TrueClass === obj || FalseClass === obj || NilClass === obj
    obj
  elsif String === obj
    obj.dup
  elsif Array === obj
    obj.map { |e| deep_dup_original_branches(e) }
  elsif Hash === obj
    duped = {}
    obj.each_pair do |key, value|
      duped[key] = deep_dup_original_branches(value)
    end
    duped
  else
    obj.dup
  end
end


obj = {
  "class" => "ActiveJob::QueueAdapters::SidekiqAdapter::JobWrapper",
  "wrapped" => "ActionMailer::MailDeliveryJob",
  "queue" => "mailers",
  "args": [
    { "job_class" => "ActionMailer::MailDeliveryJob",
      "job_id" => "2f967da1-a389-479c-9a4e-5cc059e6d65c",
      "provider_job_id" => nil,
      "queue_name" => "mailers",
      "priority" => nil,
      "arguments" => [
        "ApiMailer",
        "test_email",
        "deliver_now",
        { "args" => [1,2,3], "_aj_symbol_keys" => ["args"]}
      ],
      "executions" => 0,
      "exception_executions" => {},
      "locale" => "en",
      "timezone" => "UTC",
      "enqueued_at" => "2019-09-12T16:28:37Z"
    }],
    "retry" => true,
    "jid" => "469979df52bb9ef9f48b49e1",
    "created_at" => 1568305717.9457421,
    "enqueued_at" => 1568305717.9457731
}


Benchmark.ips do |x|
  x.report("deep_dup") do
    deep_dup(obj)
  end

  x.report("deep_dup_optimized") do
    deep_dup_optimized(obj)
  end

  x.report("deep_dup_original_branches") do
    deep_dup_original_branches(obj)
  end

  x.compare!
end
```

```
Warming up --------------------------------------
            deep_dup     5.027k i/100ms
  deep_dup_optimized     6.401k i/100ms
deep_dup_original_branches
                         5.735k i/100ms
Calculating -------------------------------------
            deep_dup     51.718k (± 1.5%) i/s -    261.404k in   5.055548s
  deep_dup_optimized     65.753k (± 1.3%) i/s -    332.852k in   5.063121s
deep_dup_original_branches
                         58.092k (± 3.8%) i/s -    292.485k in   5.043546s

Comparison:
  deep_dup_optimized:    65753.1 i/s
deep_dup_original_branches:    58091.7 i/s - 1.13x  slower
            deep_dup:    51718.3 i/s - 1.27x  slower
```